Update Mail Sender + Roles

+++ src/main/java/hexa/cloud/hexacloud/repository/VpsRepository.java
@@ -2,6 +2,6 @@ package hexa.cloud.hexacloud.repository;
 import hexa.cloud.hexacloud.model.Vps;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VpsRepository extends JpaRepository<Vps, Long> {
+public interface VpsRepository extends JpaRepository<Vps, Integer> {
 
 }
